+++ src/pipecatcloud/_utils/console_utils.py
@@ -30,10 +30,12 @@ class PipecatConsole(Console):
     def error(
             self,
             message,
+            title: Optional[str] = None,
             title_extra: Optional[str] = None,
             subtitle: Optional[str] = None):
 
-        title = f"{PANEL_TITLE_ERROR}{f' - {title_extra}' if title_extra is not None else ''}"
+        if not title:
+            title = f"{PANEL_TITLE_ERROR}{f' - {title_extra}' if title_extra is not None else ''}"
 
         self.print(
             Panel(